+++ lib/output/http_client.go
@@ -57,7 +57,10 @@ message has multiple parts the request will be sent according to
 
 // NewHTTPClient creates a new HTTPClient output type.
 func NewHTTPClient(conf Config, mgr types.Manager, log log.Modular, stats metrics.Type) (Type, error) {
-	h := writer.NewHTTPClient(conf.HTTPClient, log, stats)
+	h, err := writer.NewHTTPClient(conf.HTTPClient, log, stats)
+	if err != nil {
+		return nil, err
+	}
 	return NewWriter("http_client", h, log, stats)
 }